Cast iterable object to a list before concatenating to list
In python 3, the keys() function on a dictionary no longer returns
a list. Instead it returns an iterable object. In order to concatenate
that to a list, we must explicitly cast it to a list.

In python 2, keys() does return a list, so casting it to a list is
effectively a no-op.