Use feature detection to import modules for python 2 or 3
All of the urllib, urllib2, and urlparse modules in python 2 were moved
into different submodules of urllib in python 3. The strategy employed
here to support both versions of python is to attempt to load each of
these submodules as they would appear in python 3 and catch an
ImportError if we are in python 2. Since this makes the import block
more difficult to read, I moved these imports into a separate file
and then was able to write cleaner import statements in xplenty_api.